Fixing warnings by using jQuery and one file
