+++ Lombiq.Walkthroughs/Constants/ResourceNames.cs
@@ -5,5 +5,5 @@ public static class ResourceNames
     private const string Prefix = FeatureIds.Area + ".";
 
     public const string Shepherd = Prefix + "sheperd.js";
-
+    public const string Walkthroughs = nameof(Walkthroughs);
 }
