+++ Lombiq.Walkthroughs/Filters/WalkthroughsButtonFilter.cs
@@ -1,7 +1,6 @@
 using Microsoft.AspNetCore.Mvc.Filters;
 using OrchardCore.DisplayManagement;
 using OrchardCore.DisplayManagement.Layout;
-using OrchardCore.Mvc.Core.Utilities;
 using System;
 using System.Threading.Tasks;
