+++ Lombiq.Walkthroughs/ResourceManagementOptionsConfiguration.cs
@@ -26,22 +26,11 @@ public class ResourceManagementOptionsConfiguration : IConfigureOptions<Resource
                 "~/" + FeatureIds.Area + "/shepherd.js/js/shepherd.js");
 
         _manifest
-            .DefineScript("Walkthroughs")
+            .DefineScript(ResourceNames.Walkthroughs)
+            .SetDependencies("jQuery")
             .SetUrl(
                 "~/" + FeatureIds.Area + "/js/walkthroughs.min.js",
                 "~/" + FeatureIds.Area + "/js/walkthroughs.js");
-
-        _manifest
-            .DefineScript("ShepherdQueryParameterHelpers")
-            .SetUrl(
-                "~/" + FeatureIds.Area + "/js/shepherd-query-parameter-helpers.min.js",
-                "~/" + FeatureIds.Area + "/js/shepherd-query-parameter-helpers.js");
-
-        _manifest
-            .DefineScript("OrchardCoreAdminWalkthrough")
-            .SetUrl(
-                "~/" + FeatureIds.Area + "/js/orchard-core-admin-walkthrough.min.js",
-                "~/" + FeatureIds.Area + "/js/orchard-core-admin-walkthrough.js");
     }
 
     public void Configure(ResourceManagementOptions options) => options.ResourceManifests.Add(_manifest);
